requirements.txt
transformers
torch
torchvision
torchaudio
numpy
pandas
matplotlib
tqdm
scikit-learn
Pillow
opencv-python
easyocr
git+https://github.com/csebuetnlp/normalizer